+++ src/pages/HomePage/components/ModalContents/index.ts
@@ -1,3 +1,2 @@
 export { default as ModalContentsFriends } from './Friends/ModalContentsFriends';
-export { default as ModalContentsSetting } from './Setting/ModalContentsSetting';
 export { default as ModalContentsCategory } from './Category/ModalContentsCategory';
