+++ src/shared/components/Sidebar/assets/svgs/index.ts
@@ -1,0 +1,3 @@
+export { default as FolderIcon } from './ic_folder.svg?react';
+export { default as LogoIcon } from './ic_logo.svg?react';
+export { default as GearIcon } from './ic_gear.svg?react';
