+++ src/shared/components/Sidebar/components/ModalContents/index.ts
@@ -1,0 +1,1 @@
+export { default as ModalContentsSetting } from './Setting/ModalContentsSetting';
